AccessorBuilder should be public API
