+++ src/lib.rs
@@ -150,6 +150,9 @@
 // Add options below to allow/deny Clippy lints.
 
 // Private module with public types, they will be accessed via `opendal::Xxxx`
+mod builder;
+pub use builder::AccessorBuilder;
+
 mod operator;
 pub use operator::BatchOperator;
 pub use operator::Operator;
