+++ src/raw/mod.rs
@@ -25,7 +25,6 @@
 
 mod accessor;
 pub use accessor::Accessor;
-pub use accessor::AccessorBuilder;
 pub use accessor::AccessorCapability;
 pub use accessor::AccessorHint;
 pub use accessor::AccessorMetadata;
