+++ tests/behavior/utils.rs
@@ -22,8 +22,7 @@ use bytes::Bytes;
 use log::debug;
 use opendal::layers::LoggingLayer;
 use opendal::layers::RetryLayer;
-use opendal::raw::AccessorBuilder;
-use opendal::Operator;
+use opendal::*;
 use rand::prelude::*;
 use sha2::Digest;
 use sha2::Sha256;
